Update INC Folder

+++ src/classes/Admin.php
@@ -66,6 +66,8 @@
  		/// Get actions available for Uploaders too
 		///  la variable $this->page correspond à l'affichage de sortie 
 		/// puisque on appelle $this->page->toHTML() voir en bas du fichier		
+		
+		//Section uniquement pour les administrateurs et uploders
  		if(isset($_GET['a'])){
 	 		switch($_GET['a']){
 	 			case "Abo" 		: 	$this->page = new AdminAbout();
@@ -85,6 +87,7 @@
 											CurrentUser::$path = Settings::$photos_dir;
 										}
 									}
+
 									$json=AdminMove::move();
 	 								
 	 								if(isset($_POST['move']) && $_POST['move']=="rename"){
@@ -96,8 +99,8 @@
 											CurrentUser::$path = Settings::$photos_dir;
 										}
 									}
-		 							
 									$this->page = JS::toHTML($json);
+									$this->page = '';
 									break;
 
 				case "Del"		:	if(isset($_POST['del'])){
@@ -117,6 +120,8 @@
  		/// Get action
 		///  la variable $this->page correspond à l'affichage de sortie 
 		/// puisque on appelle $this->page->toHTML() voir en bas du fichier	
+		
+		//Section uniquement pour les administrateurs
  		if(isset($_GET['a'])){
 	 		switch($_GET['a']){
 		 		case "Sta"		:	$this->page = new AdminStats();
@@ -124,19 +129,35 @@
 
 		 		case "VTk"		:	$this->page = new AdminToken();
 		 							break;
+									
+				case "JS"			:	$this->page = new JS();
+									break;
 
-				case "DTk" 		:	if(isset($_POST['tokenkey'])){
+				case "EdA"		:	$this->page = new JSAccounts();
+									break;	
+
+				case "CTk"		:	//Create Token
+									if (isset($_POST['token'])) {
+									GuestToken::create(CurrentUser::$path);
+									}
+									$this->page = new AdminToken();
+									break;
+
+				case "DTk" 		:	//Delete Token
+									if(isset($_POST['tokenkey'])){
 										GuestToken::delete($_POST['tokenkey']);
 									}
 									$this->page = new AdminToken();
 									break;	
 
-				case "AcC"		:	if(isset($_POST['login'])){
+				case "AcC"		:	//Search Account
+									if(isset($_POST['login'])){
 										$this->page= new AdminAccount($_POST['login']);
 									}
 									break;
 
-		 		case "Acc"		:	if(isset($_POST['edit'])){
+		 		case "Acc"		:	//Edit Account
+									if(isset($_POST['edit'])){
 										Account::edit($_POST['login'],NULL,$_POST['password'],$_POST['name'],$_POST['email'],NULL,$_POST['language']);
 										$this->page= new AdminAccount($_POST['login']);
 									} else {
@@ -144,38 +165,44 @@
 									}
 									break;
 
-				case "GC"		:	if(isset($_POST['group'])){
+				case "GC"		:	//Create Group
+									if(isset($_POST['group'])){
 										Group::create($_POST['group']);
 									}
 									$this->page = new JSAccounts();
 									break;
 									
-				case "GDe"		:	if(isset($_POST['name'])){
+				case "GDe"		:	//Delete Group
+									if(isset($_POST['name'])){
 										Group::delete($_POST['name']);
 									}
 									$this->page = new JSAccounts();
 									break;									
 
-				case "AAc"		:	if($_POST){
+				case "AAc"		:	//Create Account
+									if($_POST){
 										Account::create($_POST['login'],$_POST['password'],$_POST['verif']);
 										unset($_POST);
 									}
 									$this->page= new AdminAccount($_POST['login']);
 									break;
 									
-				case "ADe"		:	if(isset($_POST['name'])){
+				case "ADe"		:	//Delete Account
+									if(isset($_POST['name'])){
 										Account::delete($_POST['name']);
 									}
 									$this->page = new JSAccounts();
 									break;									
 				
-				case "AGA"		:	$a = new Account($_POST['acc']);
+				case "AGA"		:	// Add user in group
+									$a = new Account($_POST['acc']);
 									$a->add_group($_POST['group']);
 									$a->save();
 									$this->page = new JSAccounts();
 									break;
 
-				case "AGR"		:	if($_POST){
+				case "AGR"		:	//Remove user of group
+									if($_POST){
 										$a = new Account($_POST['acc']);
 										$a->remove_group($_POST['group']);
 										$a->save();
@@ -184,52 +211,50 @@
 									$this->page = new JSAccounts();
 									break;
 
-				case "CDe"		:	if(isset($_POST['date'])){
+				case "CDe"		:	//Delete Comment
+									if(isset($_POST['date'])){
 										Comments::delete(File::r2a($_POST['path']),$_POST['date']);
 									}
 									$this->page = new Comments(CurrentUser::$path);
 									break;
-
-				//~ case "Fil"			:	$this->page = new AdminFiles();
-									//~ break;
-
-				case "JS"			:	$this->page = new JS();
-									break;
-
-				case "EdA"		:	$this->page = new JSAccounts();
-									break;
 				
-				case "GAl"		:	if(isset($_POST['path'])){
+				case "GAl"		:	//Generate Thumbs
+									if(isset($_POST['path'])){
 										Settings::gener_all(File::r2a(stripslashes($_POST['path'])));
 									}
 									$this->page = new Settings();
 									break;
 									
-				case "DAl"		:	if(isset($_POST['cleanpath'])){
+				case "DAl"		:	//Delete Thumbs
+									if(isset($_POST['cleanpath'])){
 										Settings::cleanthumbs(File::r2a(stripslashes($_POST['cleanpath'])));
 									}
 									$this->page = new Settings();
 									break;
 									
-				case "Set" 		:	if(isset($_POST['name'])){
+				case "Set" 		:	//Save Preferences settings
+									if(isset($_POST['name'])){
 										Settings::set();
 									}
 									$this->page = new Settings();
 									break;
 									
-				case "Tis"	:		if(isset($_POST['f'])){
+				case "Tis"	:		//Save Information
+									if(isset($_POST['f'])){
 										TextInfo::Save_File($_POST['f'],$_POST['title'],$_POST['author'],$_POST['contain']);
 									}
 									$this->page = new TextInfo(CurrentUser::$path);
 									break;	
 									
-				case "Tid"	:		if(isset($_POST['f'])){
+				case "Tid"	:		//Delete Information
+									if(isset($_POST['f'])){
 										TextInfo::Delete_File($_POST['f']);
 									}
 									$this->page = new TextInfo(CurrentUser::$path);
 									break;
 									
-				case "RTy"	:		if(isset($_POST['type'])){
+				case "RTy"	:		//Change permission Public/Private
+									if(isset($_POST['type'])){
 										if ($_POST['type']=='Pri'){
 											Judge::edit(CurrentUser::$path,array(),array(),true);
 										}
@@ -241,7 +266,8 @@
 									break;
 								
 
-				case "Rig"	:		if (isset($_POST['users'])|| isset($_POST['groups'])) {
+				case "Rig"	:		//Edit permissions
+									if (isset($_POST['users'])|| isset($_POST['groups'])) {
 										Judge::edit(CurrentUser::$path,$_POST['users'],$_POST['groups'],true);
 									}
 									$this->page = new Judge (CurrentUser::$path);
@@ -250,9 +276,6 @@
 				
 				
 		}
-		//~ if(!isset($this->page)){
-			//~ $this->page = new Settings();			
-		//~ }
 
 	 	/// Create menu
 	 	$this->menu = new AdminMenu();
